requirements.txt
pywin32~=310
pyserial~=3.5